feat: show identifier summary
with Luc we discussed that it might be useful to show some basic identifiers
and a link to Pubchem to give users the possibility
to verify that we retrieved the information for the correct compound.

i noticed that the labels are not optimal atm.
I'll fix this upstream in the Pubchem package.